bump arrow1 and arrow2 to latest versions (#273)
* bump arrow1

* fmt

* pin zstd-sys

* bump arrow2 as well to pin one zstd version

* unpin zstd sys

* install clang

* remove comment

+++ src/arrow2/reader_async.rs
@@ -120,8 +120,7 @@ pub async fn read_row_group(
     let mut writer = IPCStreamWriter::new(&mut output_file, options);
     writer.start(arrow_schema, None)?;
 
-    let deserializer =
-        RowGroupDeserializer::new(column_chunks, row_group_meta.num_rows() as usize, None);
+    let deserializer = RowGroupDeserializer::new(column_chunks, row_group_meta.num_rows(), None);
     for maybe_chunk in deserializer {
         let chunk = maybe_chunk?;
         writer.write(&chunk, None)?;
